requirements.txt
aiomysql==0.2.0
awswrangler==3.13.0
bcrypt==4.3.0
boto3==1.40.38
botocore==1.40.38
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
dotenv==0.9.9
google-auth==2.41.0
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-dataframe==4.0.0
httplib2==0.31.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
joblib==1.5.2
lightgbm==4.6.0
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
paramiko==2.12.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyMySQL==1.1.2
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sshtunnel==0.4.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
